rename UserPreferences to UserPrefs
